worker/requirements.txt
boto3==1.7.47
botocore==1.10.47
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.5
Pillow==5.1.0
psutil==5.4.6
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
